+++ pytest.ini
@@ -2,3 +2,6 @@
 [pytest]
 markers =
     integration: mark a test as an integration test that calls external services
+addopts = -m "not integration"
+testpaths =
+    tests
